% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_col_signature_scores.R
\docType{data}
\name{col_signature_score_char}
\alias{col_signature_score_char}
\title{Character signature score columns}
\format{
An object of class \code{character} of length 3.
}
\usage{
col_signature_score_char
}
\description{
A character vector of column names for categorical signature scores used in LundTaxR.
}
\keyword{datasets}
