% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_col_signature_scores.R
\docType{data}
\name{col_signature_score_formatted}
\alias{col_signature_score_formatted}
\title{Formatted signature score names}
\format{
An object of class \code{character} of length 19.
}
\usage{
col_signature_score_formatted
}
\description{
A named character vector mapping internal column names to formatted display names for signature scores.
}
\keyword{datasets}
