% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_col_signature_scores.R
\docType{data}
\name{col_signature_score_num}
\alias{col_signature_score_num}
\title{Numeric signature score columns}
\format{
An object of class \code{character} of length 19.
}
\usage{
col_signature_score_num
}
\description{
A character vector of column names for numeric signature scores used in LundTaxR.
}
\keyword{internal}
