% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_col_signature_scores.R
\docType{data}
\name{col_signature_score_prop}
\alias{col_signature_score_prop}
\title{Proportion signature score columns}
\format{
An object of class \code{character} of length 13.
}
\usage{
col_signature_score_prop
}
\description{
A character vector of column names for proportion signature scores used in LundTaxR.
}
\keyword{internal}
