% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_subset.R
\name{get_prediction_subset}
\alias{get_prediction_subset}
\title{Get Prediction Subsets}
\usage{
get_prediction_subset(
  these_predictions = NULL,
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  samples_rownames = FALSE
)
}
\arguments{
\item{these_predictions}{A list containing data frames and named vectors to be subset.}

\item{these_sample_ids}{A vector of sample IDs to subset the data. If NULL, sample IDs will be
extracted from \code{these_samples_metadata}.}

\item{these_samples_metadata}{A data frame containing sample metadata. If provided and
\code{these_sample_ids} is NULL, sample IDs will be extracted from this data frame.}

\item{samples_rownames}{A logical value indicating whether the sample IDs are stored in the row
names of \code{these_samples_metadata}. Default is TRUE.}
}
\value{
A list with the same structure as \code{these_predictions}, but subsetted to the specified sample IDs.
}
\description{
This function subsets the data frames and named vectors in a list based on a set of sample IDs.
}
\details{
This function is designed to subset various components of a predictions list based on a
set of sample IDs. If \code{these_sample_ids} is not provided, the function will attempt to extract
sample IDs from \code{these_samples_metadata}. The function handles different structures within the
list, including data frames and named vectors, ensuring that only the specified samples are retained.
}
\examples{
my_samples = head(sjodahl_2017_meta$sample_id)
my_meta = head(sjodahl_2017_meta)

subsetted_data <- get_prediction_subset(these_predictions = sjodahl_classes, 
                                        these_sample_ids = my_samples)

subsetted_data <- get_prediction_subset(these_predictions = sjodahl_classes, 
                                        these_samples_metadata = my_meta,
                                        samples_rownames = FALSE)

}
