% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{get_sample_metrics}
\alias{get_sample_metrics}
\title{Deprecated: Use get_subtype_metrics() instead}
\usage{
get_sample_metrics(...)
}
\arguments{
\item{...}{All arguments are passed to \code{\link{get_subtype_metrics}}}
}
\description{
\lifecycle{deprecated}
This function has been deprecated in favor of \code{\link{get_subtype_metrics}}.
}
