% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subtype_metrics.R
\name{get_subtype_metrics}
\alias{get_subtype_metrics}
\title{Get Sample Metrics}
\usage{
get_subtype_metrics(
  this_metadata = NULL,
  this_metadata_variable = NULL,
  these_predictions = NULL,
  factor_level = NULL,
  subtype_class = "5_class"
)
}
\arguments{
\item{this_metadata}{A data frame containing metadata with a \code{sample_id} column.}

\item{this_metadata_variable}{A string specifying the column name in the metadata to be summarized.}

\item{these_predictions}{A named vector of predictions with sample IDs as names.}

\item{factor_level}{The level of the factor/value in \code{this_metadata_variable} to count (optional).
If not provided, will return counts for all levels.}

\item{subtype_class}{The classification system, default is 5 class.}
}
\value{
A data frame with the number of samples and counts for each subtype.
If \code{factor_level} is specified, returns counts for that specific level.
If \code{factor_level} is NULL, returns counts for all levels.
}
\description{
This function tables a metadata column based on subtype classification.
}
\details{
The function performs the following steps:
\itemize{
\item Ensures the sample IDs are present in both metadata and predictions.
\item Checks if the desired metadata column is valid.
\item Filters metadata and predictions to include only common samples.
\item Combines metadata and predictions into a single data frame.
\item Counts the number of samples and occurrences of specified level for each subtype.
}
}
\examples{
my_predictions = classify_samples(this_data = sjodahl_2017, 
                                  gene_id = "hgnc_symbol", 
                                  impute = TRUE, 
                                  adjust = TRUE)

#count progression events (factor level 1)
get_subtype_metrics(these_predictions = my_predictions, 
                    this_metadata = sjodahl_2017_meta, 
                    this_metadata_variable = "surv_os_event", 
                    factor_level = 1)

#count specific gender (e.g., "Male")
get_subtype_metrics(these_predictions = my_predictions, 
                    this_metadata = sjodahl_2017_meta, 
                    this_metadata_variable = "gender",
                    factor_level = "Male")

}
