% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_signatures.R
\name{int_calc_signatures}
\alias{int_calc_signatures}
\title{Calculate Lund Taxonomy Signature Scores}
\usage{
int_calc_signatures(
  this_data = NULL,
  gene_id = "hgnc_symbol",
  threshold_progression = 0.58,
  threshold_grade = 0.5,
  adjust = FALSE,
  adj_factor = 5.1431,
  verbose = TRUE
)
}
\arguments{
\item{this_data}{Required parameter. Data frame or matrix with expression values.}

\item{gene_id}{Specify the type of gene identifier used in \code{this_data}. Accepted values are;
hgnc_symbol (default) or ensembl_gene_id.}

\item{threshold_progression}{Threshold to flag a sample as high/low grade, default is 0.5.}

\item{adjust}{Boolean parameter. If TRUE, the function will proceed with adjusting the scores based
on stable genes. If FALSE (default), no adjustment will be made and the original score values will be retained.}

\item{adj_factor}{Only applicable if adjust is set to TRUE. Allows users to apply a proportional
adjustment to the normalized scores, enabling finer control over the final output values.
After dividing each score by the mean expression of stable genes, the result is multiplied by this factor.}

\item{verbose}{A logical value indicating whether processing messages will be
printed or not. Default is TRUE.}
}
\value{
A list of 2, a data frame with scores for the selected variable and a data frame with
information on missing genes.
}
\description{
Wrapper function for calculating Lund Taxonomy scores.
}
\details{
This function internally calls a set of non-exported scoring functions starting with the
int_ prefix. See individual function documentation for more information on the individual function calls.
}
\examples{
#load packages
library(dplyr, multiclassPairs)

#get scores for bundled data
my_scores = int_calc_signatures(this_data = sjodahl_2017,
                                  impute = TRUE)

}
