% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_check_log2_transformation.R
\name{int_check_log2_transformation}
\alias{int_check_log2_transformation}
\title{Check if RNAseq Expression Data is Log2-Transformed}
\usage{
int_check_log2_transformation(expression_df, print_histogram = FALSE)
}
\arguments{
\item{expression_df}{A data frame containing RNAseq expression data.}

\item{print_histogram}{Set to TRUE to draw histogram with expression values, default is FALSE.}
}
\value{
A list with information about log2 transformation.
}
\description{
This function checks whether the values in an RNAseq expression data frame have been
log2-transformed. It performs multiple checks, including the range of values, the presence of zeros,
and the presence of negative values, to provide a robust assessment.
}
\details{
The function uses the following criteria to determine if the data is log2-transformed:
\itemize{
\item If the maximum value exceeds 20, the data is likely not log2-transformed.
\item If the minimum value is less than -5, the data is likely not log2-transformed.
\item If the data contains zeros, it may be log2-transformed but without pseudocounts.
\item If the range and distribution of values are consistent with log2-transformed data,
the function concludes that the data is likely log2-transformed.
}
}
\examples{
int_check_log2_transformation(expression_df = sjodahl_2017)

}
