% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deprecated_functions.R
\name{predict_LundTax2023}
\alias{predict_LundTax2023}
\title{Deprecated: Use classify_samples() instead}
\usage{
predict_LundTax2023(...)
}
\arguments{
\item{...}{All arguments are passed to \code{\link{classify_samples}}}
}
\description{
This function has been deprecated in favor of \code{\link{classify_samples}}.
}
