% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deprecated_functions.R
\name{create_deprecated_function}
\alias{create_deprecated_function}
\title{Create Deprecated Function}
\usage{
create_deprecated_function(old_name, new_name, package_name, when = "2.0.0")
}
\arguments{
\item{old_name}{Character string of deprecated function name}

\item{new_name}{Character string of replacement function name}

\item{package_name}{Character string of package name}

\item{when}{Character string indicating version when deprecation started}
}
\value{
Function that calls new function with deprecation warning
}
\description{
Create deprecated function wrapper.
}
\details{
Internal function, create deprecated function wrapper.
}
\examples{
\dontrun{
# No examples provided
}

}
