% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_order.R
\name{get_sample_order}
\alias{get_sample_order}
\title{Get Sample Order.}
\usage{
get_sample_order(expr_data = NULL, return_this = "sample_order")
}
\arguments{
\item{return_this}{Set to "late_early" to return late/early ratio. Set to "sample_order" to
return the sample order based on late/early ratio (default).}

\item{these_predictions}{Required parameter, should be the output from
\code{\link[LundTax2023Classifier:lundtax_predict_sub]{LundTax2023Classifier::lundtax_predict_sub()}}. Note, this function required the \code{lundtax_predict_sub}
to be run with \code{include_data = TRUE} argument.}
}
\value{
Rreturns the sample order.
}
\description{
Return a sample order based on the late/early ratio (proliferation score).
}
\details{
Modified version of the heatmap function, this function only returns the
order of samples to be used for downstream plotting functions. The order is dictated by the early
late ratio. and the sample order can readily be sued elsewhere, if needed.
}
\examples{
#get predictions with data
my_predictions = lundtax_predict_sub(this_data = sjodahl_2017,
                                     include_data = TRUE, 
                                     gene_id = "hgnc_symbol", 
                                     impute = TRUE, 
                                     adjust = TRUE)
                                     
#run function                                    
get_sample_order(these_predictions = my_predictions)

}
