% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{lundtax_predict_sub}
\alias{lundtax_predict_sub}
\title{Deprecated: Use classify_samples() instead}
\usage{
lundtax_predict_sub(...)
}
\arguments{
\item{...}{All arguments are passed to \code{\link{classify_samples}}}
}
\description{
\lifecycle{deprecated}
This function has been deprecated in favor of \code{\link{classify_samples}}.
}
